+++ app/src/main/java/cf/awidiyadew/sunshine_intermediate_class/model/Temp.java
@@ -30,7 +30,7 @@ public class Temp{
 		return min;
 	}
 
-	public void setMax(int max){
+	public void setMax(double max){
 		this.max = max;
 	}
